Many changes and distance functionality added
Changed:
- User is now prompted for number of days, increment, axes, and planet
to measure distance from. This avoids running the function with
seemingly arbritary parameters that make no sense to the reader/user
- Distance calculated by adding another return parameter to the plot
function, so that the points are received and can have math done
- All coordinates are put into a large 2 by 9 matrix so that values can
be accessed by distance function without having to use 9 if statements
of 9 calculations per if statement.

To do:
- Comment all code so that it is understandable to reader/user